move keys to new file
- comment out new test for now
- make new folder for new responses
