+++ emailGenerator/keys.js
@@ -1,0 +1,57 @@
+const religions = [
+  { adj: "Christian", noun: "Christian" },
+  { adj: "Muslim", noun: "Muslim" },
+  { adj: "Hindu", noun: "Hindu" },
+  { adj: "Sikh", noun: "Sikh" },
+  { adj: "Jewish", noun: "Jew" },
+  { adj: "Buddhist", noun: "Buddhist" },
+  { adj: "religious", noun: "person of faith" },
+];
+
+const questionKeys = {
+  gil6UCe4dG9T: "supportAid",
+  wKGNjgRDml1H: "motivation",
+  MRPxTl6j1QAw: "whichCountry",
+  Z4awe4sDljLR: "countryLinks",
+  EejpFBEzP9wK: "conservative",
+  IdqRPd6SUMVh: "religion",
+  vdZgYVyiLE13: "meetMp",
+  UhNb2Z5nqHtb: "meetMpDoubleCheck",
+  ghzBmQTQ2npF: "emailAddress",
+  uLPPjjg5B0Bn: "homeAddress",
+  hgdzZ05GxSAs: "postcode",
+  daZZA6TwyMP5: "name",
+};
+
+// "choices": [
+//     {
+//       "id": "bxQtiqPordqM",
+//       "label": "COVID -- as long as any people in the world have COVID, all of us are unsafe"
+//     },
+//     {
+//       "id": "P9iJPQMgOAz2",
+//       "label": "COVID -- hundreds of millions of people will be pushed into extreme poverty"
+//     },
+//     {
+//       "id": "sFjiqVGt1Qys",
+//       "label": "The Global Britain agenda and the forthcoming presidencies of the G7 and COP26 make it particularly important that we show moral leadership"
+//     },
+//     {
+//       "id": "03hZJbLCuqUh",
+//       "label": "Foreign aid improves national security and reduces immigration"
+//     },
+//     {
+//       "id": "eyqNtdDrwY38",
+//       "label": "If we can find money for military spending and nuclear weapons we can find money to help the poor"
+//     },
+//     {
+//       "id": "CUydnMGQ0p6p",
+//       "label": "Foreign aid is morally right because it reduces suffering and injustice"
+//     },
+//     {
+//       "id": "CR1RaEx9YifJ",
+//       "label": "Money goes further overseas, so we can do so much more good with our money there"
+//     }
+//   ]
+
+module.exports = { religions, questionKeys };
